+++ R/inequalityMeasure.R
@@ -2,8 +2,9 @@
 #'
 #' @description Computes Allison and Foster inequality measure of a given variable taking into account weights.
 #'
-#' @param X is a data vector
+#' @param X is a data vector (numeric or ordered factor)
 #' @param W is a vector of weights
+#' @param norm (logical). If TRUE (default) then index is divided by a maximum possible value which is a difference between maximum and minimum of X
 #'
 #' @importFrom stats aggregate
 #'
@@ -11,50 +12,49 @@
 #'
 #' @rdname Allison_and_Foster
 #'
-#' @details Let \eqn{m} be the median category and \eqn{p_i} be a share of \eqn{i}-th category. The following index was proposed by Allison and Foster (2004):
-#'
-#' @details
+#' @details Let \eqn{c=(c_{1},...,c_{n})} be the vector of categories in increasing order, \eqn{m} be the median category and \eqn{p_i} be a share of \eqn{i}-th category. The following index was proposed by Allison and Foster (2004):
 #' \deqn{AF =  \frac{\sum_{i=m}^n c_{i} p_{i} }{\sum_{i=m}^n p_{i}} - \frac{\sum_{i=1}^{m-1} c_{i} p_{i}}{\sum_{i=1}^{m-1} p_{i}}}
+#' Note that above formula is valid only for numerical values. Thus, in order to compute AF for ordered factor, X is converted to numerical variable.
 #'
-#' @references Atkinson A. B.: (1970) On the measurement of inequality
-#' @references Coulter P. B.: (1989) Measuring Inequality ISBN 0-8133-7726-9
-#' @references Cowell F. A.: (2011) Measuring Inequality (Third Edition) ISBN 978-0-19-959403-0
 #' @references Allison R. A., Foster J E.: (2004) Measuring health inequality using qualitative data, Journal of Health Economics
-#' @references Kobus M., Miłoś P.: (2011) Inequality decomposition by population subgroups for ordinal data, Working Papers, No. 24
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' AF(X)
 #' AF(X,W)
 #'
 #' data(Well_being)
-#' # Allison and Foster index for health assessment with sample weights:
+#' # Allison and Foster index for health assessment with sample weights
 #' X=Well_being$V11
 #' W=Well_being$Weight
 #' AF(X,W)
 #'
 #'
 #' @export
-AF=function(X,W=rep(1,length(X)))
+AF=function(X,W=rep(1,length(X)),norm=TRUE)
 {
-  W=W[is.na(W)==FALSE]
-  W=aggregate(W,by=list(X),FUN=sum)
-  Fx=(cumsum(W$x)/sum(W$x))
-  SW=cumsum(W$x)
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) & !is.ordered(X))return('X must be numeric or ordered factor')
+  if(length(unique(X))==1)return(0)
+  X=as.numeric(X)
+  tab=aggregate(W,by=list(X),FUN=sum)
+  Fx=(cumsum(tab$x)/sum(tab$x))
+  SW=cumsum(tab$x)
   min=min(X)
   max=max(X)
   a=which(Fx<0.5)
   if(length(a)!=0)
   {A=max(a)
-  b=floor(sum(W$x)/2)-SW[max(a)]
-  c=W$x[max(a)+1]-b }else{A=0
-  b=floor(sum(W$x)/2)
-  c=W$x[1]-b}
-  AF=((sum(W$Group.1[-c(a,A+1)]*W$x[-c(a,A+1)])+W$Group.1[A+1]*c)/(sum(W$x[-c(a,A+1)])+c))-((sum(W$Group.1[a]*W$x[a])+W$Group.1[A+1]*b)/(sum(W$x[a])+b))
-  AFNorm=AF/(max-min)
-  return(AFNorm)
+  b=floor(sum(tab$x)/2)-SW[max(a)]
+  c=tab$x[max(a)+1]-b }else{A=0
+  b=floor(sum(tab$x)/2)
+  c=tab$x[1]-b}
+  AF=((sum(tab$Group.1[-c(a,A+1)]*tab$x[-c(a,A+1)])+tab$Group.1[A+1]*c)/(sum(tab$x[-c(a,A+1)])+c))-((sum(tab$Group.1[a]*tab$x[a])+tab$Group.1[A+1]*b)/(sum(tab$x[a])+b))
+  if(norm){return(AF/(max-min))}else{return(AF)}
 }
 
 
@@ -79,24 +79,28 @@ AF=function(X,W=rep(1,length(X)))
 #' for \eqn{\epsilon=1}.
 #'
 #' @references Atkinson A. B.: (1970) On the measurement of inequality, Journal of Economic Theory
-#' @references Coulter P. B.: (1989) Measuring Inequality ISBN 0-8133-7726-9
-#' @references Allison P.D.: (1978) Measures of Inequality, American Sociological Review
+#'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Atkinson(X)
 #' Atkinson(X,W)
 #'
 #' data(Tourism)
-#' # Atkinson index for Total expenditure with sample weights:
-#' X= Tourism$`Total expenditure`
+#' # Atkinson index for Total expenditure with sample weights
+#' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Atkinson(X,W)
 #'
 #' @export
 Atkinson=function(X,W=rep(1,length(X)),e=1)
 {
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(0)
   if(e==1){A=1-prod(X^(W/sum(W)))/ (sum(W*X)/sum(W))}else{
     A=1-((1/sum(W)*sum(W*(X^(1-e))))^(1/(1-e)))/(sum(W*X)/sum(W))}
   return(A)
@@ -110,9 +114,7 @@ Atkinson=function(X,W=rep(1,length(X)),e=1)
 #' @param X is a data vector
 #' @param W is a vector of weights
 #' @param parameter is a entropy parameter
-#' @param na.rm logical, should missing values (NAs) be removed prior to computations? If set to FALSE the computations yield NA
 #'
-#' @importFrom stats na.omit
 #'
 #' @return The value of generalized entropy index
 #'
@@ -127,28 +129,27 @@ Atkinson=function(X,W=rep(1,length(X)),e=1)
 #' @references Pielou E.C.: (1966) The measurement of diversity in different types of biological collections. Journal of Theoretical Biology
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Entropy(X)
 #' Entropy(X,W)
 #'
 #' data(Tourism)
-#' # Generalized entropy index for Total expenditure with sample weights:
-#' X= Tourism$`Total expenditure`
+#' # Generalized entropy index for Total expenditure with sample weights
+#' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Entropy(X,W)
 #'
 #'
 #' @export
-Entropy=function (X,W=rep(1,length(X)), parameter = 0.5, na.rm = TRUE)
+Entropy=function (X,W=rep(1,length(X)), parameter = 0.5)
 {
-  if (!na.rm && any(is.na(X)))
-    return(NA_real_)
-  X <- as.numeric(na.omit(X))
-  X <- as.numeric(X)
-  if (is.null(parameter))
-    parameter <- 0.5
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(0)
   if (parameter == 0)
     e <- Theil_L(X,W)
   else if (parameter == 1)
@@ -169,7 +170,6 @@ Entropy=function (X,W=rep(1,length(X)), parameter = 0.5, na.rm = TRUE)
 #' @param X is a data vector
 #' @param W is a vector of weights
 #' @param parameter is a Kolm parameter
-#' @param na.rm logical, should missing values (NAs) be removed prior to computations? If set to FALSE the computations yield NA
 #' @param scale method of data scaling (None, Normalization, Unitarization, Standardization)
 #'
 #' @importFrom stats na.omit
@@ -190,7 +190,7 @@ Entropy=function (X,W=rep(1,length(X)), parameter = 0.5, na.rm = TRUE)
 #'
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Kolm(X)
@@ -209,24 +209,21 @@ Entropy=function (X,W=rep(1,length(X)), parameter = 0.5, na.rm = TRUE)
 #' Kolm(10*X,scale ="Standardization")
 #'
 #' @export
-Kolm=function (X,W=rep(1,length(X)), parameter = 1, na.rm = TRUE, scale = "None")
+Kolm=function (X,W=rep(1,length(X)), parameter = 1, scale = "None")
 {
-  if (!na.rm && any(is.na(X)))
-    return(NA_real_)
-  X <- as.numeric(na.omit(X))
-  X <- as.numeric(X)
-
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(0)
   if(scale=="Standardization"){X=(X-mean(X))/sd(X)}
   if(scale=="Unitarization"){X=(X-min(X))/(max(X)-min(X))}
   if(scale=="Normalization"){X=X/sqrt(sum(X^2))}
-
-  if (is.null(parameter))
-    parameter <- 1
   W <- W/sum(W)
   KM <- parameter * (sum(W*X) - X)
   KM <- sum(W*(exp(KM)))
   KM <- (1/parameter) * log(KM)
-  KM
+  return(KM)
 }
 
 
@@ -236,7 +233,6 @@ Kolm=function (X,W=rep(1,length(X)), parameter = 1, na.rm = TRUE, scale = "None"
 #'
 #' @param X is a data vector
 #' @param W is a vector of weights
-#' @param na.rm logical, should missing values (NAs) be removed prior to computations? If set to FALSE the computations yield NA
 #'
 #' @importFrom stats na.omit
 #'
@@ -252,28 +248,30 @@ Kolm=function (X,W=rep(1,length(X)), parameter = 1, na.rm = TRUE, scale = "None"
 #' @references Eliazar I. I., Sokolov I. M.: (2010) Measuring statistical heterogeneity: The Pietra index
 #' @references Costa R. N., Pérez-Duarte S.: (2019) Not all inequality measures were created equal, Statistics Paper Series, No 31
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' RicciSchutz(X)
 #' RicciSchutz(X,W)
 #'
 #' data(Tourism)
-#' #Ricci and Schutz index for Total expenditure with sample weights:
-#' X= Tourism$`Total expenditure`
+#' #Ricci and Schutz index for Total expenditure with sample weights
+#' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' RicciSchutz(X,W)
 #'
 #'
 #' @export
-RicciSchutz=function (X,W=rep(1,length(X)), na.rm = TRUE)
+RicciSchutz=function (X,W=rep(1,length(X)))
 {
-  if (!na.rm && any(is.na(X)))
-    return(NA_real_)
-  X <- as.numeric(na.omit(X))
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(0)
   d <- abs(X - (sum(W*X)/sum(W)))
   d <- (sum(W*d)/sum(W))/(2 * (sum(W*X)/sum(W)))
-  d
+  return(d)
 }
 
 
@@ -286,7 +284,6 @@ RicciSchutz=function (X,W=rep(1,length(X)), na.rm = TRUE)
 #' @param X is a data vector
 #' @param W is a vector of weights
 #' @param square logical, argument of the function CoefVar, for details see below
-#' @param na.rm logical, should missing values (NAs) be removed prior to computations? If set to FALSE the computations yield NA
 #'
 #' @importFrom stats na.omit
 #'
@@ -303,30 +300,30 @@ RicciSchutz=function (X,W=rep(1,length(X)), na.rm = TRUE)
 #' @references Coulter P. B.: (1989) Measuring Inequality ISBN 0-8133-7726-9
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' CoefVar(X)
 #' CoefVar(X,W)
 #'
 #' data(Tourism)
-#' #Coefficient of variation for Total expenditure with sample weights:
-#' X= Tourism$`Total expenditure`
+#' #Coefficient of variation for Total expenditure with sample weights
+#' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' CoefVar(X,W)
 #'
 #'
 #' @export
-CoefVar=function (X,W=rep(1,length(X)), square = FALSE, na.rm = TRUE)
+CoefVar=function (X,W=rep(1,length(X)), square = FALSE)
 {
-  if (!na.rm && any(is.na(X)))
-    return(NA_real_)
-  x <- as.numeric(na.omit(X))
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(0)
   w.m <- sum(W*X)/sum(W)
   V <- sqrt(sum(W*(X-w.m)^2)/sum(W))/w.m
-  if (square)
-    V <- V^2
-  V
+  if(square){return(V^2)}else{return(V)}
 }
 
 
@@ -344,26 +341,34 @@ CoefVar=function (X,W=rep(1,length(X)), square = FALSE, na.rm = TRUE)
 #' @details Gini coefficient is given by:
 #'          \deqn{G = \frac{ \sum_{i=1}^n \sum_{j=1}^n \mid x_{i} - x_{j} \mid}{2n^{2} \overline{x}}}
 #'
-#' @references Coulter P. B.: (1989) Measuring Inequality ISBN 0-8133-7726-9
 #' @references Dixon P. M., Weiner, J., Mitchell-Olds, T., and Woodley, R.: (1987) Bootstrapping the Gini Coefficient of Inequality. Ecology , Volume 68 (5)
 #' @references Firebaugh G.: (1999) Empirics of World Income Inequality, American Journal of Sociology
 #' @references Deininger K.; Squire L.: (1996) A New Data Set Measuring Income Inequality, The World Bank Economic Review, Vol. 10, No. 3
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Gini(X)
 #' Gini(X,W)
 #'
 #' data(Tourism)
-#' #Gini coefficient for Total expenditure with sample weights:
-#' X= Tourism$`Total expenditure`
+#' #Gini coefficient for Total expenditure with sample weights
+#' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Gini(X,W)
 #'
 #'
 #' @export
-Gini=function(X,W=rep(1,length(X))){sum(abs(matrix(X,length(X),length(X),TRUE)-matrix(X,length(X),length(X),FALSE))*matrix(W,length(W),length(W),TRUE)*matrix(W,length(W),length(W),FALSE)/(2*sum(W)^2*(sum(W*X)/sum(W))))}
+Gini=function(X,W=rep(1,length(X)))
+{
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(0)
+  G=sum(abs(matrix(X,length(X),length(X),TRUE)-matrix(X,length(X),length(X),FALSE))*matrix(W,length(W),length(W),TRUE)*matrix(W,length(W),length(W),FALSE)/(2*sum(W)^2*(sum(W*X)/sum(W))))
+  return(G)
+}
 
 
 
@@ -382,26 +387,34 @@ Gini=function(X,W=rep(1,length(X))){sum(abs(matrix(X,length(X),length(X),TRUE)-m
 #' \deqn{H={\frac {1}{2}}{\frac {\sum_{i}|x_{i}-{\overline{x}}|}{\sum_{i}x_{i}}}}
 #'
 #'
-#' @references Coulter P. B.: (1989) Measuring Inequality ISBN 0-8133-7726-9
 #' @references Hoover E. M. Jr.: (1936) The Measurement of Industrial Localization, The Review of Economics and Statistics, 18
 #' @references Hoover E. M. Jr.: (1984) An Introduction to Regional Economics, ISBN 0-07-554440-7
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Hoover(X)
 #' Hoover(X,W)
 #'
 #' data(Tourism)
-#' #Hoover index for Total expenditure with sample weights:
+#' #Hoover index for Total expenditure with sample weights
 #' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Hoover(X,W)
 #'
 #'
 #' @export
-Hoover=function(X,W=rep(1,length(X))){(1/2)*(sum(W*abs(X - (sum(W*X)/sum(W))))/sum(X*W))}
+Hoover=function(X,W=rep(1,length(X)))
+{
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(0)
+  H=(1/2)*(sum(W*abs(X - (sum(W*X)/sum(W))))/sum(X*W))
+  return(H)
+}
 
 
 #' @title Leti index
@@ -426,15 +439,15 @@ Hoover=function(X,W=rep(1,length(X))){(1/2)*(sum(W*abs(X - (sum(W*X)/sum(W))))/s
 #'
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Leti(X)
 #' Leti(X,W)
 #'
 #' data(Tourism)
-#' #Leti index for Total expenditure with sample weights:
-#' X= Tourism$`Total expenditure`
+#' #Leti index for Total expenditure with sample weights
+#' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Leti(X,W)
 #'
@@ -442,10 +455,13 @@ Hoover=function(X,W=rep(1,length(X))){(1/2)*(sum(W*abs(X - (sum(W*X)/sum(W))))/s
 #' @export
 Leti=function(X,W=rep(1,length(X)),norm=T)
 {
-  W=W[is.na(W)==FALSE]
-  W=aggregate(W,by=list(sort(X,decreasing = F)),FUN=sum)
-  W=W$x
-  Fx=(cumsum(W)/sum(W))
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) & !is.ordered(X))return('X must be numeric or ordered factor')
+  if(length(unique(X))==1)return(0)
+  tab=aggregate(W,by=list(X),FUN=sum)
+  Fx=(cumsum(tab$x)/sum(tab$x))
   Leti=2*sum(Fx*(1-Fx))
   if(norm){Leti=Leti*2/(length(Fx)-1)}
   return(Leti)
@@ -481,53 +497,55 @@ Leti=function(X,W=rep(1,length(X)),norm=T)
 #'
 #'
 #'
-#' @references Coulter P. B.: (1989) Measuring Inequality ISBN 0-8133-7726-9
 #' @references Jenkins S. P. and P. J. Lambert: (1997) Three ‘I’s of Poverty Curves, with an Analysis of U.K. Poverty Trends
 #' @references Cowell F. A.: (2000) Measurement of Inequality, Handbook of Income Distribution
 #' @references Cowell F. A., Flachaire E.: (2017) Inequality with Ordinal Data
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Jenkins(X)
 #' Jenkins(X,W)
 #'
 #' data(Tourism)
-#' #Jenkins, Cowell and Flachaire coefficients for Total expenditure with sample weights:
+#' #Jenkins, Cowell and Flachaire coefficients for Total expenditure with sample weights
 #' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Jenkins(X,W)
 #'
 #'
 #' @export
-Jenkins=function(X,W=rep(1,length(X)), alfa=0.8) #dane to wektor (konkretne pytanie), 0<= alfa < 1
+Jenkins=function(X,W=rep(1,length(X)), alfa=0.8)
 {
-  dane=as.vector(X)
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(0)
   tab=aggregate(W,by=list(X),FUN=sum)
 
   yy=as.data.frame(tab)
-  yy$fk=yy$x/sum(yy$x) #odsetek
-  yy$Fk=cumsum(yy$x)/sum(yy$x) #dystrybuanta
+  yy$fk=yy$x/sum(yy$x) #PDF
+  yy$Fk=cumsum(yy$x)/sum(yy$x) #CDF
 
-  dejta=as.data.frame(tab$Group.1)
-  dejta = merge(dejta,yy[,c("Group.1","fk","Fk")],by.y="Group.1",by.x = "tab$Group.1",all.x = TRUE)
-
+  df=as.data.frame(tab$Group.1)
+  df = merge(df,yy[,c("Group.1","fk","Fk")],by.y="Group.1",by.x = "tab$Group.1",all.x = TRUE)
 
   if(alfa!=0)
-  {Cowell_and_Flachaire=1/(alfa*(alfa-1))*(1/nrow(dejta)*sum((dejta$Fk)^alfa)-1)
+  {Cowell_and_Flachaire=1/(alfa*(alfa-1))*(1/nrow(df)*sum((df$Fk)^alfa)-1)
   }else{
-    Cowell_and_Flachaire=-1/nrow(dejta)*sum(log(dejta$Fk))}
+    Cowell_and_Flachaire=-1/nrow(df)*sum(log(df$Fk))}
 
-  P_I=seq(1:nrow(dejta))/nrow(dejta)
+  P_I=seq(1:nrow(df))/nrow(df)
 
-  GL=1/nrow(dejta)*cumsum(dejta$Fk)
+  GL=1/nrow(df)*cumsum(df$Fk)
 
-  dejta$P_I=P_I
-  dejta$GL=GL
+  df$P_I=P_I
+  df$GL=GL
 
-  GL_J=aggregate(dejta$GL,by=list(dejta$`tab$Group.1`),FUN=max)
-  P_I_J=aggregate(dejta$P_I,by=list(dejta$`tab$Group.1`),FUN=max)
+  GL_J=aggregate(df$GL,by=list(df$`tab$Group.1`),FUN=max)
+  P_I_J=aggregate(df$P_I,by=list(df$`tab$Group.1`),FUN=max)
 
   JJ=rbind(c(0,0,0),cbind(GL_J,P_I_J$x))
   colnames(JJ)=c("Group","GL_J","PI_J")
@@ -547,7 +565,7 @@ Jenkins=function(X,W=rep(1,length(X)), alfa=0.8) #dane to wektor (konkretne pyta
 #'
 #' @description Palma proportion - originally the ratio of the total income of the 10% richest people to the 40% poorest people.
 #'
-#' @param X is a data vector
+#' @param X is a data vector (numeric or ordered factor)
 #' @param W is a vector of weights
 #'
 #' @importFrom stats aggregate
@@ -564,14 +582,14 @@ Jenkins=function(X,W=rep(1,length(X)), alfa=0.8) #dane to wektor (konkretne pyta
 #' @references Cobham A., Sumner A.: (2013) Putting the Gini Back in the Bottle? 'The Palma' as a Policy-Relevant Measure of Inequality
 #' @references Palma J. G.: (2011) Homogeneous middles vs. heterogeneous tails, and the end of the ‘Inverted-U’: the share of the rich is what it’s all about
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Palma(X)
 #' Palma(X,W)
 #'
 #' data(Tourism)
-#' #Palma index for Total expenditure with sample weights:
+#' #Palma index for Total expenditure with sample weights
 #' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Palma(X,W)
@@ -581,20 +599,28 @@ Jenkins=function(X,W=rep(1,length(X)), alfa=0.8) #dane to wektor (konkretne pyta
 #' @export
 Palma=function(X,W=rep(1,length(X)))
 {
-  W=W[is.na(W)==F]
-  W=aggregate(W,by=list(X),FUN=sum)
-  Fx=(cumsum(W$x)/sum(W$x))
-  SW=cumsum(W$x)
-  a=which(Fx<0.4)
-  if(length(a)!=0){A=max(a)
-  b=floor(sum(W$x)*0.4)-SW[max(a)]}else{A=0
-  b=floor(sum(W$x)*0.4)}
-  c=which(Fx<0.9)
-  if(length(c)!=0){C=max(c)
-  d=W$x[max(c)+1]-(floor(sum(W$x)*0.9)-SW[max(c)])}else{C=0
-  d=W$x[1]-(floor(sum(W$x)*0.9))}
-  Palma=(sum(W$Group.1[-c(c,C+1)]*W$x[-c(c,C+1)])+W$Group.1[C+1]*d)/(sum(W$Group.1[a]*W$x[a])+W$Group.1[A+1]*b)
-
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(1)
+  W=W[order(X)];X=X[order(X)]
+  Fx=(cumsum(W)/sum(W))
+  nominator=sum(X[which(Fx>=0.9)])
+  denominator=ifelse(min(Fx)<=0.4,sum(X[which(Fx<=0.4)]),X[1])
+  # tab=aggregate(W,by=list(X),FUN=sum)
+  # Fx=(cumsum(tab$x)/sum(tab$x))
+  # SW=cumsum(tab$x)
+  # a=which(Fx<0.4)
+  # if(length(a)!=0){A=max(a)
+  # b=floor(sum(tab$x)*0.4)-SW[max(a)]}else{A=0
+  # b=floor(sum(tab$x)*0.4)}
+  # c=which(Fx<0.9)
+  # if(length(c)!=0){C=max(c)
+  # d=tab$x[max(c)+1]-(floor(sum(tab$x)*0.9)-SW[max(c)])}else{C=0
+  # d=tab$x[1]-(floor(sum(tab$x)*0.9))}
+  # Palma=(sum(tab$Group.1[-c(c,C+1)]*tab$x[-c(c,C+1)])+tab$Group.1[C+1]*d)/(sum(tab$Group.1[a]*tab$x[a])+tab$Group.1[A+1]*b)
+  Palma=nominator/denominator
   return(Palma)
 }
 
@@ -602,7 +628,7 @@ Palma=function(X,W=rep(1,length(X)))
 #'
 #' @description 20:20 ratio - originally the ratio of the total income of the 20% richest people to the 20% poorest people.
 #'
-#' @param X is a data vector
+#' @param X is a data vector (numeric or ordered factor)
 #' @param W is a vector of weights
 #'
 #' @importFrom stats aggregate
@@ -623,14 +649,14 @@ Palma=function(X,W=rep(1,length(X)))
 #'
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Prop20_20(X)
 #' Prop20_20(X,W)
 #'
 #' data(Tourism)
-#' #Prop20_20 proportion for Total expenditure with sample weights:
+#' #Prop20_20 proportion for Total expenditure with sample weights
 #' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Prop20_20(X,W)
@@ -639,21 +665,30 @@ Palma=function(X,W=rep(1,length(X)))
 #' @export
 Prop20_20=function(X,W=rep(1,length(X)))
 {
-  W=W[is.na(W)==FALSE]
-  W=aggregate(W,by=list(X),FUN=sum)
-  Fx=(cumsum(W$x)/sum(W$x))
-  SW=cumsum(W$x)
-  a=which(Fx<0.2)
-  a=which(Fx<0.2)
-  if(length(a)!=0){A=max(a)
-  b=floor(sum(W$x)*0.2)-SW[max(a)]}else{A=0
-  b=floor(sum(W$x)*0.2)}
-  c=which(Fx<0.8)
-  if(length(c)!=0){C=max(c)
-  d=W$x[max(c)+1]-(floor(sum(W$x)*0.8)-SW[max(c)])}else{C=0
-  d=W$x[1]-(floor(sum(W$x)*0.8))}
-  Prop20_20=(sum(W$Group.1[-c(c,C+1)]*W$x[-c(c,C+1)])+W$Group.1[C+1]*d)/(sum(W$Group.1[a]*W$x[a])+W$Group.1[A+1]*b)
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  if(length(unique(X))==1)return(0)
+  W=W[order(X)];X=X[order(X)]
+  Fx=(cumsum(W)/sum(W))
+  nominator=sum(X[which(Fx>=0.8)])
+  denominator=ifelse(min(Fx)<=0.2,sum(X[which(Fx<=0.2)]),X[1])
 
+  # W=aggregate(W,by=list(X),FUN=sum)
+  # Fx=(cumsum(tab$x)/sum(tab$x))
+  # SW=cumsum(tab$x)
+  # a=which(Fx<0.2)
+  # a=which(Fx<0.2)
+  # if(length(a)!=0){A=max(a)
+  # b=floor(sum(tab$x)*0.2)-SW[max(a)]}else{A=0
+  # b=floor(sum(tab$x)*0.2)}
+  # c=which(Fx<0.8)
+  # if(length(c)!=0){C=max(c)
+  # d=tab$x[max(c)+1]-(floor(sum(tab$x)*0.8)-SW[max(c)])}else{C=0
+  # d=tab$x[1]-(floor(sum(tab$x)*0.8))}
+  # Prop20_20=(sum(tab$Group.1[-c(c,C+1)]*tab$x[-c(c,C+1)])+tab$Group.1[C+1]*d)/(sum(tab$Group.1[a]*tab$x[a])+tab$Group.1[A+1]*b)
+  Prop20_20=nominator/denominator
   return(Prop20_20)
 }
 
@@ -677,14 +712,14 @@ Prop20_20=function(X,W=rep(1,length(X)))
 #' @references OECD: (2020) Regions and Cities at a Glance 2020, Chapter: Indexes and estimation techniques
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Theil_L(X)
 #' Theil_L(X,W)
 #'
 #' data(Tourism)
-#' # Theil L coefficient for Total expenditure with sample weights:
+#' # Theil L coefficient for Total expenditure with sample weights
 #' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Theil_L(X,W)
@@ -692,7 +727,14 @@ Prop20_20=function(X,W=rep(1,length(X)))
 #'
 #'
 #' @export
-Theil_L=function(X,W=rep(1,length(X))){1/sum(W)*sum(W*log((sum(W*X)/sum(W))/X))}
+Theil_L=function(X,W=rep(1,length(X)))
+{
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  return(1/sum(W)*sum(W*log((sum(W*X)/sum(W))/X)))
+}
 
 
 
@@ -716,14 +758,14 @@ Theil_L=function(X,W=rep(1,length(X))){1/sum(W)*sum(W*log((sum(W*X)/sum(W))/X))}
 #' @references OECD: (2020) Regions and Cities at a Glance 2020, Chapter: Indexes and estimation techniques
 #'
 #' @examples
-#' # Compare weighted and unweighted result.
+#' # Compare weighted and unweighted result
 #' X=1:10
 #' W=1:10
 #' Theil_T(X)
 #' Theil_T(X,W)
 #'
 #' data(Tourism)
-#' # Theil T coefficient for Total expenditure with sample weights:
+#' # Theil T coefficient for Total expenditure with sample weights
 #' X=Tourism$`Total expenditure`
 #' W=Tourism$`Sample weight`
 #' Theil_T(X,W)
@@ -731,7 +773,221 @@ Theil_L=function(X,W=rep(1,length(X))){1/sum(W)*sum(W*log((sum(W*X)/sum(W))/X))}
 #'
 #'
 #' @export
-Theil_T=function(X,W=rep(1,length(X))){1/sum(W)*sum(W*(X/(sum(W*X)/sum(W))*log(X/(sum(W*X)/sum(W)))))}
+Theil_T=function(X,W=rep(1,length(X)))
+{
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) | !is.numeric(W))return('X and W must be numeric')
+  return(1/sum(W)*sum(W*(X/(sum(W*X)/sum(W))*log(X/(sum(W*X)/sum(W))))))
+}
+
+#' @title Abul Naga and Yalcin index
+#'
+#' @description Computes Abul Naga and Yalcin inequality measure of a given variable taking into account weights.
+#'
+#' @param X is a data vector (numeric or ordered factor)
+#' @param W is a vector of weights
+#' @param a is a positive parameter. See more in details
+#' @param b is a positive parameter. See more in details
+#'
+#' @importFrom stats aggregate
+#'
+#' @return The value of Abul Naga and Yalcin coefficient.
+#'
+#' @rdname Abul_Naga_and_Yalcin
+#'
+#' @details Let \eqn{m} be the median category, \eqn{n} be the number of categories and \eqn{P_{i}} be the cumulative distribution of \eqn{i}-th category.
+#' The following index with respect to the parameters a and b was proposed by Abul Naga and Yalcin (2008):
+#' \deqn{I=\frac{a\sum_{i<m}^{n}P_{i}-b\sum_{i\geq m}^{n}P_{i}+b(n+1-m)}{0.5(a(m-1)+b(n-m))}}
+#'
+#' @references Ramses H. Abul Naga and Tarik Yalcin: (2008) Inequality Measurement for ordered response health data, Journal of Health Economics 27(6);
+#'
+#' @examples
+#' # Compare weighted and unweighted result
+#' X=1:10
+#' W=1:10
+#' AN_Y(X)
+#' AN_Y(X,W)
+#'
+#' data(Well_being)
+#' # Abul Naga and Yalcin index for health assessment with sample weights
+#' X=Well_being$V1
+#' W=Well_being$Weight
+#' AN_Y(X,W)
+#'
+#'
+#' @export
+
+AN_Y=function(X,W=rep(1,length(X)),a=1,b=1)
+{
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) & !is.ordered(X))return('X must be numeric or ordered factor')
+  if(length(unique(X))==1)return(AB_Y=0)
+  tab=aggregate(W,by=list(X),FUN=sum)
+  #dystrybuanta
+  Fx=(cumsum(tab$x)/sum(tab$x))
+  #mediana
+  m_b=max(c(1,which(Fx<0.5)),na.rm=T) #ponizej mediany
+  m=min(which(Fx>=0.5))
+  sum_bm=sum(Fx[1:m_b])
+  sum_am=sum(Fx)-sum(Fx[1:m_b])
+  n=nrow(tab)
+  Abul_Naga_Yelcin=(a*sum_bm-b*sum_am+b*(n+1-m))/(((a*(m-1))+b*(n-m))/2)
+
+  return(Abul_Naga_Yelcin)
+}
+
+#' @title Apouey index
+#'
+#' @description Computes Apouey inequality measure of a given variable taking into account weights.
+#'
+#' @param X is a data vector (numeric or ordered factor)
+#' @param W is a vector of weights
+#' @param a is a positive parameter. See more in details
+#' @param b is a real parameter. See more in details
+#'
+#' @importFrom stats aggregate
+#'
+#' @return The value of Apouey coefficient.
+#'
+#' @rdname Apouey
+#'
+#' @details Let \eqn{m} be the median category, \eqn{n} will be the number of categories and \eqn{P_i} be the cumulative distribution of \eqn{i}-th category. The following index was proposed by Apouey (2007):
+#' \deqn{I = \alpha(\sum_{i\geq m}^{n}P_{i}-\sum_{i<m}^{n}P_{i}+m-\frac{n}{2}-1)+\beta}
+#' where \eqn{\alpha} and \eqn{\beta} are given parameters with default values \eqn{\alpha=\frac{2}{1-n}} and \eqn{\beta=\frac{n}{n-1}}.
+#' @references Apouey B.: (2007) Measuring health polarization with self-assessed health data, Health Economics 16; 875-894.
+#'
+#' @examples
+#' # Compare weighted and unweighted result
+#' X=1:10
+#' W=1:10
+#' Apouey(X,a=2,b=2)
+#' Apouey(X,W,a=2,b=2)
+#'
+#' data(Well_being)
+#' # Apouey index for health assessment with sample weights
+#' X=Well_being$V1
+#' W=Well_being$Weight
+#' Apouey(X,W,a=2,b=2)
+#'
+#'
+#' @export
+
+Apouey=function(X,W=rep(1,length(X)),a=2/(1-length(W[!is.na(W) & !is.na(X)])),b=length(W[!is.na(W) & !is.na(X)])/(length(W[!is.na(W) & !is.na(X)])-1))
+{
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) & !is.ordered(X))return('X must be numeric or ordered factor')
+  if(length(unique(X))==1)return(Apouey=0)
+  tab=aggregate(W,by=list(X),FUN=sum)
+  Fx=(cumsum(tab$x)/sum(tab$x))
+  m_b=max(c(1,which(Fx<0.5)),na.rm=T)
+  m=min(which(Fx>=0.5))
+  sum_bm=sum(Fx[1:m_b])
+  sum_am=sum(Fx)-sum(Fx[1:m_b])
+  n=nrow(tab)
+  Apouey=a*(sum_am-sum_bm+m-(n/2)-1)+b
+
+  return(Apouey)
+}
+
+
+#' @title Blair and Lacy index
+#'
+#' @description Computes Blair and Lacy inequality measure of a given variable taking into account weights.
+#'
+#' @param X is a data vector (numeric or ordered factor)
+#' @param W is a vector of weights
+#' @param withsqrt if TRUE function returns index given by BL2, elsewhere by BL (default). See more in details.
+#'
+#' @importFrom stats aggregate
+#'
+#' @return The value of Blair and Lacy coefficient.
+#'
+#' @rdname Blair_Lacy
+#'
+#' @details Let \eqn{m} be the median category, \eqn{n} be the number of categories and \eqn{P_i} be the cumulative distribution of \eqn{i}-th category.
+#' The indices of Blair and Lacy (2000) are the following:
+#' \deqn{BL =  1-\frac{\sum_{i=1}^{n-1}(P_{i}-0.5)^2}{\frac{n-1}{4}}}
+#' \deqn{BL2 =  1-\left(\frac{\sum_{i=1}^{n-1}(P_{i}-0.5)^2}{\frac{n-1}{4}}\right)^{\frac{1}{2}}}
+#'
+#' @references Blair J, Lacy M G. (2000): Statistics of ordinal variation, Sociological Methods and Research 28(251);251-280.
+#'
+#' @examples
+#' # Compare weighted and unweighted result
+#' X=1:10
+#' W=1:10
+#' BL(X)
+#' BL(X,W)
+#'
+#' data(Well_being)
+#' # Blair and Lacy index for health assessment with sample weights
+#' X=Well_being$V1
+#' W=Well_being$Weight
+#' BL(X,W)
+#'
+#'
+#' @export
+#Blair and Lacy Index
+BL=function(X,W=rep(1,length(X)),withsqrt=FALSE)
+{
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) & !is.ordered(X))return('X must be numeric or ordered factor')
+  if(length(unique(X))==1)return(BL=0)
+  tab=aggregate(W,by=list(X),FUN=sum)
+  Fx=(cumsum(tab$x)/sum(tab$x))
+  n=nrow(tab)
+  BL=1-(sum(Fx[1:(n-1)])-0.5)^2/(n-1)/4
+  BL2=1-sqrt((sum(Fx[1:(n-1)])-0.5)^2/(n-1)/4)
+  if(withsqrt){return(BL2)}else{return(BL)}
+}
+
+#' @title Median of ordered factor or numeric
+#'
+#' @description Computes median of ordered factor or numeric variable taking into account weights.
+#'
+#' @param X is a data vector (numeric or ordered factor)
+#' @param W is a vector of weights
+#'
+#' @importFrom stats aggregate
+#'
+#' @return The median category (number or label) of ordered factor.
+#'
+#' @rdname medianf
+#'
+#' @details Calculates median based on cumulative distribution. Tailored for ordered factors.
+#'
+#'
+#' @examples
+#' # Compare weighted and unweighted result
+#' X=factor(c('H','H','M','M','L','L'),levels = c('L','M','H'),ordered = TRUE)
+#' W=c(2,2,3,3,8,8)
+#' medianf(X)
+#' medianf(X,W)
+#'
+#'
+#'
+#' @export
+#Median category
+medianf=function(X,W=rep(1,length(X)))
+{
+  ind=which(!is.na(W) & !is.na(X))
+  if(length(ind)==0)return('Input with NAs only')
+  W=W[ind];X=X[ind]
+  if(!is.numeric(X) & !is.ordered(X))return('X must be numeric or ordered factor')
+  tab=aggregate(W,by=list(X),FUN=sum)
+  Fx=(cumsum(tab$x)/sum(tab$x))
+  m=min(which(Fx>=0.5))
+  return(medianf=tab$Group.1[m])
+}
+
+
 #' @title Sample survey on trips
 #' @description Data from sample survey on trips conducted in Polish households.
 #' @docType data